the great double conversion, epsiode 5 changed write to wb, works better maybe? but its slower floored bound box GOT RID OF THE BLACK PIXEL

+++ proj3/rast.h
@@ -40,7 +40,7 @@ class rast
 
   void iLight(iobj* tri);
   void objN(iobj* tri);
-  void intColour(float& triU,float& triV,iobj* tri);
+  void intColour(double& triU,double& triV,iobj* tri);
   
   void writeImg(); //write colours from img array
     
@@ -51,38 +51,38 @@ class rast
   SlVector3 m_up;
   int m_angle;
   int m_dim;
-  float m_hither;
+  double m_hither;
   unsigned char m_background[3];
-  float m_backgroundF[3]; //float version for accurate calculation
+  double m_backgroundF[3]; //double version for accurate calculation
   unsigned char m_colour[3]; //my windows computer -> G B R colour for some reason (correct in gl)
-  float m_colourF[3];
+  double m_colourF[3];
   
   /*--calculated values for mathyness--*/
   int m_psize; //pixel grid size
   SlVector3 m_w;
   SlVector3 m_u;
   SlVector3 m_v;
-  float m_d;
-  float m_m;
+  double m_d;
+  double m_m;
 
   /*--data arrays--*/
   SlVector3* m_ppointsV; //pixel points vector
                          //size should be psize
 
   iobj* m_light; //lights list
-  float m_maxLight; //max lights (not really data but goes with m lights)
+  double m_maxLight; //max lights (not really data but goes with m lights)
   iobj* m_adata; //alldata list
-  float** m_img; //image data
+  double** m_img; //image data
 
   /*--M matrix temps--*/
-  float m_Mtemp[4];
+  double m_Mtemp[4];
 
   /*--calcboundbox stuff--*/
-  float m_boundBox[4]; //bounding box [min x,min y,max x,max y]
+  double m_boundBox[4]; //bounding box [min x,min y,max x,max y]
 
   /*--fillP temps--*/
   SlVector3 m_baryT[3]; //temp bary vectors
-  float m_baryTF[5]; //temp dot products
+  double m_baryTF[5]; //temp dot products
 
   /*--ilight temps--*/
   SlVector3 m_lray;
